null チェックを追加

+++ ThGameMgr.Ex/MainWindow.xaml.cs
@@ -2673,24 +2673,28 @@ namespace ThGameMgr.Ex
             if (ReplayFilesDataGrid.Items.Count > 0
                 && ReplayFilesDataGrid.SelectedIndex >= 0)
             {
-                ReplayFileInfo selectedReplayFileInfo = ReplayFilesDataGrid.SelectedItem as ReplayFileInfo;
-                string replayFileName = selectedReplayFileInfo.FileName;
-
-                MessageBoxResult result = MessageBox.Show(
-                    this, $"'{replayFileName}' を削除してもよろしいですか。", "リプレイファイルの削除",
-                    MessageBoxButton.YesNo, MessageBoxImage.Information);
-                if (result == MessageBoxResult.Yes)
+                ReplayFileInfo? selectedReplayFileInfo = ReplayFilesDataGrid.SelectedItem as ReplayFileInfo;
+                if (selectedReplayFileInfo != null
+                    && selectedReplayFileInfo.FileName != null)
                 {
-                    try
-                    {
-                        ReplayFile.Delete(this.GameId, replayFileName);
+                    string replayFileName = selectedReplayFileInfo.FileName;
 
-                        GetReplayFiles();
-                    }
-                    catch (Exception ex)
+                    MessageBoxResult result = MessageBox.Show(
+                        this, $"'{replayFileName}' を削除してもよろしいですか。", "リプレイファイルの削除",
+                        MessageBoxButton.YesNo, MessageBoxImage.Information);
+                    if (result == MessageBoxResult.Yes)
                     {
-                        MessageBox.Show(this, $"リプレイファイルの削除に失敗しました。\n{ex.Message}", "エラー",
-                            MessageBoxButton.OK, MessageBoxImage.Error);
+                        try
+                        {
+                            ReplayFile.Delete(this.GameId, replayFileName);
+
+                            GetReplayFiles();
+                        }
+                        catch (Exception ex)
+                        {
+                            MessageBox.Show(this, $"リプレイファイルの削除に失敗しました。\n{ex.Message}", "エラー",
+                                MessageBoxButton.OK, MessageBoxImage.Error);
+                        }
                     }
                 }
             }
